requirements.txt
asgiref==3.7.2
Django==4.2.7
django-ckeditor==6.7.0
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-shortuuidfield==0.1.3
django-taggit==5.0.1
Pillow==10.1.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
python-decouple==3.8